pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "lmlib"
version = "0.0.1"
description = "Logistics Management Library for Maritime Industry"
authors = [
    {name = "Aryant PS", email = "aryant@computationelewetenschapsgroep.com"},
]
dependencies = [
    "requests",
    'importlib-metadata; python_version<"3.10"',
    'pydantic>=2.10',
    'pytest',
    'pytest-mock',
    'pytest-cov',
    'azure-digitaltwins-core>=1.2',
    'azure-identity>=1.19',
    'sqlglot'
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
]

[tool.setuptools.packages.find]
where = ["src"]